+++ DiagnoseServerLag/GroupReport.cs
@@ -230,7 +230,7 @@ namespace DiagnoseServerLag
                             + "ping_round_trip,quality_local,quality_remote,in_bytes_sec,out_bytes_sec,"
                             + "send_queue_bytes,send_rate_bytes_sec,zdos,instances,zdos_sent_sec,zdos_recv_sec,"
                             + "change_queue,peers,cpu_ms_per_sec,cpu_measured,gc0,gc1,gc2,collections,"
-                            + "heap_bytes,working_set_bytes,owned_ai,nearby_ai");
+                            + "heap_bytes,working_set_bytes,owned_ai,nearby_ai,feed_ms");
 
                 foreach (var s in window) sb.AppendLine(Row("server", s, c));
 
@@ -284,6 +284,7 @@ namespace DiagnoseServerLag
                 Machine.Collections(s).ToString(c),
                 s.HeapBytes.ToString(c), s.WorkingSetBytes.ToString(c),
                 s.OwnedAI.ToString(c), s.NearbyAI.ToString(c),
+                s.FeedMs.ToString("0", c),
             });
 
         private static string Iso(long ticks) =>